+++ src/vs/workbench/contrib/positronPlots/browser/components/plotsContainer.tsx
@@ -23,7 +23,7 @@ import { WebviewPlotClient } from '../webviewPlotClient.js';
 import { PlotClientInstance } from '../../../../services/languageRuntime/common/languageRuntimePlotClient.js';
 import { DarkFilter, IPositronPlotClient, IPositronPlotsService } from '../../../../services/positronPlots/common/positronPlots.js';
 import { StaticPlotClient } from '../../../../services/positronPlots/common/staticPlotClient.js';
-import { RenderFormat } from 'positron';
+import { RenderFormat } from '../../../../api/common/positron/extHostTypes.positron.js';
 
 /**
  * PlotContainerProps interface.
